+++ dnnlib/util.py
@@ -10,29 +10,28 @@
 
 import ctypes
 import fnmatch
+import glob
+import hashlib
+import html
 import importlib
 import inspect
-import numpy as np
-import os
-import shutil
-import sys
-import time
-import types
 import io
+import os
 import pickle
 import re
-import requests
-import html
-import hashlib
-import glob
+import shutil
+import sys
 import tempfile
+import time
+import types
 import urllib
 import urllib.request
 import uuid
-
 from distutils.util import strtobool
 from typing import Any, List, Tuple, Union
 
+import numpy as np
+import requests
 
 # Util classes
 # ------------------------------------------------------------------------------------------